(cms): Fixes #94
Simplify logic of setting and updating position of card header on resize
and rerender within custom hook;
